feat: Refactor random sellers retrieval to RealtorService and enhance functionality
